% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{add_rule}
\alias{add_rule}
\title{Add a new rule to a non-terminal}
\usage{
add_rule(grammar, non_terminal, rule_str, prob = NULL)
}
\arguments{
\item{grammar}{Grammar structure}

\item{non_terminal}{Name of the non-terminal}

\item{rule_str}{String representation of the rule}

\item{prob}{Probability for the new rule (if NULL, probabilities will be redistributed)}
}
\value{
Updated grammar
}
\description{
Add a new rule to a non-terminal
}
