% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{analyze_grammar_usage}
\alias{analyze_grammar_usage}
\title{Calculate statistics about grammar usage}
\usage{
analyze_grammar_usage(grammar, population)
}
\arguments{
\item{grammar}{Grammar structure}

\item{population}{Population of individuals}
}
\value{
Data frame with statistics
}
\description{
Calculate statistics about grammar usage
}
