% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_min_depths}
\alias{calculate_min_depths}
\title{Find minimum tree depth required for each non-terminal}
\usage{
calculate_min_depths(grammar)
}
\arguments{
\item{grammar}{Grammar structure}
}
\value{
Named vector with minimum depths
}
\description{
Find minimum tree depth required for each non-terminal
}
