% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{check_reachability}
\alias{check_reachability}
\title{Check which non-terminals are reachable from the start symbol}
\usage{
check_reachability(grammar)
}
\arguments{
\item{grammar}{Grammar structure to check}
}
\value{
Character vector of reachable non-terminals
}
\description{
Check which non-terminals are reachable from the start symbol
}
\keyword{internal}
