% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{create_grammar}
\alias{create_grammar}
\title{Create a new Grammar instance}
\usage{
create_grammar(grammar_path = NULL)
}
\arguments{
\item{grammar_path}{Character path to grammar file}
}
\value{
A new Grammar object
}
\description{
Create a new Grammar instance
}
