% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{crossover}
\alias{crossover}
\title{Crossover operator for PSGE}
\usage{
crossover(parent1, parent2)
}
\arguments{
\item{parent1}{First parent individual}

\item{parent2}{Second parent individual}
}
\value{
New individual created by crossover
}
\description{
Crossover operator for PSGE
}
