% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{evaluate_population}
\alias{evaluate_population}
\title{Evaluate all individuals in a population}
\usage{
evaluate_population(
  population,
  grammar,
  eval_func,
  max_depth = 17,
  parallel = FALSE,
  n_workers = 2,
  future_strategy = "multisession"
)
}
\arguments{
\item{population}{List of individuals}

\item{grammar}{PCFG grammar structure}

\item{eval_func}{Fitness evaluation function}

\item{max_depth}{Maximum tree depth}

\item{parallel}{Boolean indicating whether to use parallel evaluation}

\item{n_workers}{Number of workers for parallel processing (default: 2)}

\item{future_strategy}{Future strategy to use (default: "multisession")}
}
\value{
Evaluated population with fitness values
}
\description{
Evaluate all individuals in a population
}
\details{
Note that we use progressr progress bars with this function:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(progressr)
handlers(global = TRUE)
handlers("progress")

with_progress(\{
  population <- evaluate_population(
    population = initial_pop,
    grammar = my_grammar,
    eval_func = my_fitness_function,
    parallel = TRUE
  )
\})
}\if{html}{\out{</div>}}
}
