% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{extract_tokens}
\alias{extract_tokens}
\title{Extract tokens (terminals and non-terminals) from a production string}
\usage{
extract_tokens(production)
}
\arguments{
\item{production}{String containing a production rule}
}
\value{
List with symbols and their types
}
\description{
Extract tokens (terminals and non-terminals) from a production string
}
\keyword{internal}
