% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{generate_random_individual}
\alias{generate_random_individual}
\title{Generate a random individual for PSGE}
\usage{
generate_random_individual(grammar, max_depth = 17)
}
\arguments{
\item{grammar}{Grammar structure from read_pcfg_grammar()}

\item{max_depth}{Maximum depth for the generated tree}
}
\value{
Genotype representation as a list of real values, one list per non-terminal
}
\description{
Generate a random individual for PSGE
}
