% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{get_non_recursive_rules}
\alias{get_non_recursive_rules}
\title{Get non-recursive rules for a non-terminal}
\usage{
get_non_recursive_rules(grammar, non_terminal)
}
\arguments{
\item{grammar}{Grammar structure}

\item{non_terminal}{Name of the non-terminal}
}
\value{
List of non-recursive rules
}
\description{
Get non-recursive rules for a non-terminal
}
