% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{get_rules}
\alias{get_rules}
\title{Get rules for a specific non-terminal}
\usage{
get_rules(grammar, non_terminal)
}
\arguments{
\item{grammar}{Grammar structure}

\item{non_terminal}{Name of the non-terminal}
}
\value{
List of rules for the non-terminal
}
\description{
Get rules for a specific non-terminal
}
