% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{grammar_from_json}
\alias{grammar_from_json}
\title{Load grammar from JSON}
\usage{
grammar_from_json(json_str, is_file = FALSE)
}
\arguments{
\item{json_str}{JSON string or file path}

\item{is_file}{Whether json_str is a file path}
}
\value{
Grammar structure
}
\description{
Load grammar from JSON
}
