% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{grammar_to_bnf}
\alias{grammar_to_bnf}
\title{Export a grammar back to BNF format}
\usage{
grammar_to_bnf(grammar, file = NULL, include_probs = TRUE)
}
\arguments{
\item{grammar}{Grammar structure}

\item{file}{Output file path (optional)}

\item{include_probs}{Whether to include probabilities as comments}
}
\value{
BNF string if file is NULL, otherwise writes to file
}
\description{
Export a grammar back to BNF format
}
