% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{grammar_to_json}
\alias{grammar_to_json}
\title{Serialize grammar to JSON}
\usage{
grammar_to_json(grammar, file = NULL)
}
\arguments{
\item{grammar}{Grammar structure to serialize}

\item{file}{Output file path (optional)}
}
\value{
JSON string if file is NULL, otherwise writes to file
}
\description{
Serialize grammar to JSON
}
