% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{identify_recursive_rules}
\alias{identify_recursive_rules}
\title{Identify recursive and non-recursive rules}
\usage{
identify_recursive_rules(grammar)
}
\arguments{
\item{grammar}{Grammar structure}
}
\value{
Updated grammar with recursive_rules flags
}
\description{
Identify recursive and non-recursive rules
}
