% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{individual_to_string}
\alias{individual_to_string}
\title{Convert individual to string representation}
\usage{
individual_to_string(individual, include_genotype = FALSE)
}
\arguments{
\item{individual}{Individual to convert}

\item{include_genotype}{Whether to include genotype details}
}
\value{
String representation of individual
}
\description{
Convert individual to string representation
}
