% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_genotype}
\alias{map_genotype}
\title{Map a genotype to a phenotype using PSGE principles}
\usage{
map_genotype(grammar, genotype, max_depth = 17)
}
\arguments{
\item{grammar}{Grammar structure from read_pcfg_grammar()}

\item{genotype}{Genotype representation (list of real values per non-terminal)}

\item{max_depth}{Maximum tree depth allowed}
}
\value{
List containing phenotype string and mapping information
}
\description{
Map a genotype to a phenotype using PSGE principles
}
