% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{measure_diversity}
\alias{measure_diversity}
\title{Measure diversity in a population}
\usage{
measure_diversity(population)
}
\arguments{
\item{population}{List of individuals}
}
\value{
List of diversity metrics
}
\description{
Measure diversity in a population
}
