% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{mutate}
\alias{mutate}
\title{Mutation operator for PSGE}
\usage{
mutate(individual, mutation_rate = 0.1)
}
\arguments{
\item{individual}{Individual to mutate}

\item{mutation_rate}{Probability of mutating each codon}
}
\value{
Mutated individual
}
\description{
Mutation operator for PSGE
}
