% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{new_grammar}
\alias{new_grammar}
\title{Create an empty grammar structure}
\usage{
new_grammar(start)
}
\arguments{
\item{start}{Name of the start symbol}
}
\value{
Empty grammar structure
}
\description{
Create an empty grammar structure
}
