% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{OrderedUniqueSet}
\alias{OrderedUniqueSet}
\title{Initialize the OrderedUniqueSet}
\description{
Initialize the OrderedUniqueSet

Initialize the OrderedUniqueSet
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OrderedUniqueSet-new}{\code{OrderedUniqueSet$new()}}
\item \href{#method-OrderedUniqueSet-add}{\code{OrderedUniqueSet$add()}}
\item \href{#method-OrderedUniqueSet-has}{\code{OrderedUniqueSet$has()}}
\item \href{#method-OrderedUniqueSet-delete}{\code{OrderedUniqueSet$delete()}}
\item \href{#method-OrderedUniqueSet-values}{\code{OrderedUniqueSet$values()}}
\item \href{#method-OrderedUniqueSet-size}{\code{OrderedUniqueSet$size()}}
\item \href{#method-OrderedUniqueSet-print}{\code{OrderedUniqueSet$print()}}
\item \href{#method-OrderedUniqueSet-clone}{\code{OrderedUniqueSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedUniqueSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedUniqueSet-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedUniqueSet$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An OrderedUniqueSet object.
Add an element to the OrderedUniqueSet.

If the element is already present, it is not added again (uniqueness).
Insertion order is preserved.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedUniqueSet-add"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedUniqueSet-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedUniqueSet$add(element)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{element}}{The element to add.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the element was added (i.e., it was new), FALSE otherwise.
Check if an element is in the OrderedUniqueSet.

O(1) lookup using the environment.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedUniqueSet-has"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedUniqueSet-has}{}}}
\subsection{Method \code{has()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedUniqueSet$has(element)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{element}}{The element to check for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the element is in the set, FALSE otherwise.
Remove an element from the OrderedUniqueSet.

Maintains insertion order of remaining elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedUniqueSet-delete"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedUniqueSet-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedUniqueSet$delete(element)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{element}}{The element to remove.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the element was removed, FALSE if it was not found.
Get all elements in the order they were inserted.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedUniqueSet-values"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedUniqueSet-values}{}}}
\subsection{Method \code{values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedUniqueSet$values()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of elements in insertion order.
Get the number of elements in the OrderedUniqueSet.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedUniqueSet-size"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedUniqueSet-size}{}}}
\subsection{Method \code{size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedUniqueSet$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The size of the set.
Print the OrderedUniqueSet (for debugging/inspection)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedUniqueSet-print"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedUniqueSet-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedUniqueSet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OrderedUniqueSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OrderedUniqueSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OrderedUniqueSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
