% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_evolution}
\alias{plot_evolution}
\title{Visualize the evolution history}
\usage{
plot_evolution(history, title = "PSGE Evolution")
}
\arguments{
\item{history}{Evolution history data frame}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Visualize the evolution history
}
