% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_recursive_mapping}
\alias{rcpp_recursive_mapping}
\title{Recursive Mapping in PSGE}
\usage{
rcpp_recursive_mapping(
  grammar,
  pcfg,
  non_terminals,
  mapping_rules,
  positions_to_map,
  start_symbol,
  max_depth
)
}
\arguments{
\item{grammar}{List containing grammar structure}

\item{pcfg}{NumericMatrix of PCFG probabilities}

\item{non_terminals}{CharacterVector of non-terminal symbols}

\item{mapping_rules}{List of current mapping rules}

\item{positions_to_map}{IntegerVector of current mapping positions}

\item{start_symbol}{List specifying initial symbol}

\item{max_depth}{Integer maximum derivation depth}
}
\value{
List containing:
\describe{
\item{output}{CharacterVector of mapped phenotype}
\item{max_depth}{Integer maximum depth reached}
\item{mapping_rules}{List updated mapping rules}
}
}
\description{
Performs genotype to phenotype mapping using PSGE's probabilistic approach.
}
