% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{recursive_individual_creation}
\alias{recursive_individual_creation}
\title{Recursively create a random individual}
\usage{
recursive_individual_creation(
  grammar,
  genotype,
  symbol,
  current_depth,
  max_depth
)
}
\arguments{
\item{grammar}{Grammar structure}

\item{genotype}{Current genotype being built}

\item{symbol}{Current symbol to expand}

\item{current_depth}{Current tree depth}

\item{max_depth}{Maximum allowed tree depth}
}
\value{
Updated genotype and reached depth
}
\description{
Recursively create a random individual
}
\keyword{internal}
