% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpsge-package.R
\docType{package}
\name{Rpsge-package}
\alias{Rpsge-package}
\alias{Rpsge}
\title{Probabilistic Structured Grammatical Evolution in R}
\description{
Implementation of Probabilistic Structured Grammatical Evolution (PSGE) in R.
PSGE combines Structured Grammatical Evolution with probabilistic grammar-guided
genetic programming to evolve solutions to various problems.
}
\details{
This package provides:
\itemize{
\item A complete implementation of the PSGE algorithm
\item Grammar parsing and validation
\item Probabilistic mapping mechanisms
\item Evolutionary operators
\item High-performance C++ core through Rcpp
}
}
\author{
\strong{Maintainer}: Stéfan Janse van Rensburg \email{stefanj@mandala.ac.za} (\href{https://orcid.org/0000-0002-0749-2277}{ORCID})

}
