% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{safe_rule_selection}
\alias{safe_rule_selection}
\title{Safely select a rule based on codon value and rule probabilities}
\usage{
safe_rule_selection(rules, codon)
}
\arguments{
\item{rules}{List of rules with probabilities}

\item{codon}{Probabilistic value for selection (0-1)}
}
\value{
Selected rule index
}
\description{
Safely select a rule based on codon value and rule probabilities
}
\keyword{internal}
