% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sample_phenotypes}
\alias{sample_phenotypes}
\title{Generate a representative sample of phenotypes from a grammar}
\usage{
sample_phenotypes(grammar, n = 10, max_depth = 10)
}
\arguments{
\item{grammar}{Grammar structure}

\item{n}{Number of samples to generate}

\item{max_depth}{Maximum tree depth}
}
\value{
List of generated phenotypes
}
\description{
Generate a representative sample of phenotypes from a grammar
}
