% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{set_probabilities}
\alias{set_probabilities}
\title{Set probabilities for rules of a non-terminal}
\usage{
set_probabilities(grammar, non_terminal, probabilities)
}
\arguments{
\item{grammar}{Grammar structure}

\item{non_terminal}{Name of the non-terminal}

\item{probabilities}{Vector of probabilities}
}
\value{
Updated grammar
}
\description{
Set probabilities for rules of a non-terminal
}
