% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sort_by_fitness}
\alias{sort_by_fitness}
\title{Sort population by fitness}
\usage{
sort_by_fitness(population, minimize = TRUE)
}
\arguments{
\item{population}{List of individuals with fitness values}

\item{minimize}{Whether to minimize (TRUE) or maximize (FALSE) fitness}
}
\value{
Sorted population
}
\description{
Sort population by fitness
}
\keyword{internal}
