% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{tournament_selection}
\alias{tournament_selection}
\title{Tournament selection for PSGE}
\usage{
tournament_selection(population, tournament_size = 3)
}
\arguments{
\item{population}{List of evaluated individuals}

\item{tournament_size}{Number of individuals to include in each tournament}
}
\value{
Selected individual
}
\description{
Tournament selection for PSGE
}
