% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{UnorderedUniqueSet}
\alias{UnorderedUniqueSet}
\title{Initialize the UnorderedUniqueSet}
\description{
Initialize the UnorderedUniqueSet

Initialize the UnorderedUniqueSet
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnorderedUniqueSet-new}{\code{UnorderedUniqueSet$new()}}
\item \href{#method-UnorderedUniqueSet-add}{\code{UnorderedUniqueSet$add()}}
\item \href{#method-UnorderedUniqueSet-has}{\code{UnorderedUniqueSet$has()}}
\item \href{#method-UnorderedUniqueSet-delete}{\code{UnorderedUniqueSet$delete()}}
\item \href{#method-UnorderedUniqueSet-values}{\code{UnorderedUniqueSet$values()}}
\item \href{#method-UnorderedUniqueSet-size}{\code{UnorderedUniqueSet$size()}}
\item \href{#method-UnorderedUniqueSet-print}{\code{UnorderedUniqueSet$print()}}
\item \href{#method-UnorderedUniqueSet-clone}{\code{UnorderedUniqueSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnorderedUniqueSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnorderedUniqueSet-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnorderedUniqueSet$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An UnorderedUniqueSet object.
Add an element to the UnorderedUniqueSet.

If the element is already present, it is not added again (uniqueness).
Order is NOT preserved (it's an unordered set).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnorderedUniqueSet-add"></a>}}
\if{latex}{\out{\hypertarget{method-UnorderedUniqueSet-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnorderedUniqueSet$add(element)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{element}}{The element to add.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the element was added (i.e., it was new), FALSE otherwise.
Check if an element is in the UnorderedUniqueSet.

O(1) lookup using the environment.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnorderedUniqueSet-has"></a>}}
\if{latex}{\out{\hypertarget{method-UnorderedUniqueSet-has}{}}}
\subsection{Method \code{has()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnorderedUniqueSet$has(element)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{element}}{The element to check for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the element is in the set, FALSE otherwise.
Remove an element from the UnorderedUniqueSet.

Order is not maintained, so removal is simpler.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnorderedUniqueSet-delete"></a>}}
\if{latex}{\out{\hypertarget{method-UnorderedUniqueSet-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnorderedUniqueSet$delete(element)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{element}}{The element to remove.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the element was removed, FALSE if it was not found.
Get all elements in the UnorderedUniqueSet (order is not guaranteed).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnorderedUniqueSet-values"></a>}}
\if{latex}{\out{\hypertarget{method-UnorderedUniqueSet-values}{}}}
\subsection{Method \code{values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnorderedUniqueSet$values()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of elements (order may vary).
Get the number of elements in the UnorderedUniqueSet.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnorderedUniqueSet-size"></a>}}
\if{latex}{\out{\hypertarget{method-UnorderedUniqueSet-size}{}}}
\subsection{Method \code{size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnorderedUniqueSet$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The size of the set.
Print the UnorderedUniqueSet (for debugging/inspection)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnorderedUniqueSet-print"></a>}}
\if{latex}{\out{\hypertarget{method-UnorderedUniqueSet-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnorderedUniqueSet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnorderedUniqueSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnorderedUniqueSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnorderedUniqueSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
