% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_genotype}
\alias{validate_genotype}
\title{Validate that a genotype conforms to grammar structure}
\usage{
validate_genotype(genotype, grammar)
}
\arguments{
\item{genotype}{Genotype to validate}

\item{grammar}{Grammar structure}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Validate that a genotype conforms to grammar structure
}
