% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{evaluate_population}
\alias{evaluate_population}
\title{Evaluate all individuals in a population}
\usage{
evaluate_population(population, grammar, eval_func, max_depth = 17)
}
\arguments{
\item{population}{List of individuals}

\item{grammar}{PCFG grammar structure}

\item{eval_func}{Fitness evaluation function}

\item{max_depth}{Maximum tree depth}
}
\value{
Evaluated population with fitness values
}
\description{
Evaluate all individuals in a population
}
