% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{initialize_population}
\alias{initialize_population}
\title{Initialize a population of random individuals}
\usage{
initialize_population(grammar, pop_size, max_depth = 17)
}
\arguments{
\item{grammar}{PCFG grammar structure}

\item{pop_size}{Population size}

\item{max_depth}{Maximum tree depth}
}
\value{
List of individuals with genotypes
}
\description{
Initialize a population of random individuals
}
