% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{psge}
\alias{psge}
\title{Complete PSGE Pipeline}
\usage{
psge(grammar_path, fitness_fn, ...)
}
\arguments{
\item{grammar_path}{Path to BNF grammar file}

\item{fitness_fn}{Fitness function}

\item{...}{Additional parameters passed to setup_psge()}
}
\value{
Best solution found
}
\description{
Complete PSGE Pipeline
}
