% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psge.R
\name{psge}
\alias{psge}
\title{Complete PSGE pipeline}
\usage{
psge(grammar_file, fitness_fn, ...)
}
\arguments{
\item{grammar_file}{Path to BNF grammar file}

\item{fitness_fn}{Fitness evaluation function}

\item{...}{Additional parameters passed to run_psge}
}
\value{
Best solution and final grammar
}
\description{
Complete PSGE pipeline
}
