% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_psge}
\alias{run_psge}
\title{Run PSGE Evolution}
\usage{
run_psge(psge_instance, verbose = TRUE)
}
\arguments{
\item{psge_instance}{PSGE instance from setup_psge()}

\item{verbose}{Print progress (default TRUE)}
}
\value{
Best solution found
}
\description{
Run PSGE Evolution
}
